refactor: hooks imports
- made adjustment in `Image` to avoid require cycle

SVA-306
