+++ src/components/Image.js
@@ -7,7 +7,7 @@ import { Image as RNEImage } from 'react-native-elements';
 import { consts, colors } from '../config';
 import { imageHeight, imageWidth } from '../helpers';
 import { SettingsContext } from '../SettingsProvider';
-import { useInterval } from '../hooks';
+import { useInterval } from '../hooks/TimeHooks';
 
 import { ImageMessage } from './ImageMessage';
 import { ImageRights } from './ImageRights';
