+++ src/components/widgets/SurveyWidget.tsx
@@ -4,8 +4,7 @@ import { useQuery } from 'react-apollo';
 
 import { consts, Icon, normalize, texts } from '../../config';
 import { graphqlFetchPolicy } from '../../helpers';
-import { useRefreshTime } from '../../hooks';
-import { useHomeRefresh } from '../../hooks/HomeRefresh';
+import { useHomeRefresh, useRefreshTime } from '../../hooks';
 import { NetworkContext } from '../../NetworkProvider';
 import { SURVEYS } from '../../queries/survey';
 import { Survey, WidgetProps } from '../../types';
